requirements.txt
Flask==1.0.2
gunicorn==19.8.1
numpy==1.14.0
pytest==3.2.5
python-dateutil==2.6.1
python-dotenv==0.8.2
python-twitter==3.4.1
requests==2.18.4
requests-file==1.4.3
requests-oauthlib==0.8.0
twython==3.6.0
bokeh==0.13.0
pandas==0.23.4
virtualenv==15.2.0
virtualenv-clone==0.3.0
w3lib==1.18.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
zope.interface==4.4.3
